requirements.txt
# Dependencies for publish-blog-post.py
# Image processing (resize, optimize)
Pillow>=12.0.0

# YAML front matter parsing
PyYAML>=6.0.3
